requirements.txt
pandas>=2.2
numpy>=1.26
scipy>=1.11
pyarrow>=16.0
python-dotenv>=1.0
